requirements.txt
discord.py
pillow
pet-pet-gif